#956 Phase 3: extract cos/admission.rs from tx.rs
Per docs/pr/956-phase3-admission/plan.md (v6, post Codex round 1-4
+ Gemini round 1-2 reviews), extract the admission policy
subsystem from tx.rs into cos/admission.rs.

Move list (~475 LOC of code, ~600 LOC including doc-comment
blocks):

- Functions:
  * apply_cos_admission_ecn_policy (per-flow / aggregate ECN)
  * cos_queue_flow_share_limit (rate-aware per-flow share cap)
  * cos_flow_aware_buffer_limit (active-flow aware buffer cap)
  * bdp_floor_bytes (per-flow BDP floor; pub for tests)
  * apply_cos_queue_flow_fair_promotion (whole-runtime entry)
  * promote_cos_queue_flow_fair (per-queue, file-private)

- Constants:
  * COS_FLOW_FAIR_MIN_SHARE_BYTES + const_assert
  * COS_FLOW_FAIR_MAX_QUEUE_DELAY_NS + const_assert
  * COS_ECN_MARK_THRESHOLD_NUM/_DEN + 2 const_asserts
  * RTT_TARGET_NS, SHARED_EXACT_BURST_HEADROOM (file-private)

Visibility model:
- Production-callable items get pub(in crate::afxdp), re-exported
  via pub(super) use from cos/mod.rs.
- Test-referenced items (bdp_floor_bytes + 4 constants) are
  always-on pub(in crate::afxdp) for simplicity.
- promote_cos_queue_flow_fair stays file-private (only caller is
  apply_cos_queue_flow_fair_promotion).
- Promotion doc block at original tx.rs:5401-5467 (separated by
  V-min code) reattached to promote_cos_queue_flow_fair.

Deferred to later phases:
- account_cos_queue_flow_enqueue / _dequeue stay in tx.rs through
  Phase 4. They handle MQFQ ordering bookkeeping and V-min slot
  participation; admission gates only READ flow_bucket_bytes /
  active_flow_buckets, never call account_*. Splitting them now
  would split MQFQ and V-min invariants across files. Phase 5
  (cos/queue_ops.rs) moves them cohesively with selection / pop /
  publish.
- COS_MIN_BURST_BYTES stays in tx.rs (91 occurrences); visibility
  bumped from private to pub(in crate::afxdp) so admission.rs can
  reach it. Phase 4 / 5 should consolidate to types.rs or
  cos/mod.rs once the back-reference is no longer worth the cost.

Stale-text cleanup:
- cos/ecn.rs:1-12 — past-tense the "Phase 3 will move" prose.
- cos/flow_hash.rs:15 — past-tense "Phase 3 will import".
- cos/mod.rs:1-5 — bump phase header to current state.
- tx.rs cos-imports comment block at 3479+ — list Phase 3 as
  shipped; admission constants moved.

Net change: -485 lines from tx.rs (17683 -> 17198 lines).

Tests: cargo build clean; cargo test --bins -> 865 passed,
0 failed (the same 2 ignored fixture tests as Phase 2).

Reviewers reached PLAN-READY:
- Codex rounds 1-4 (gpt-5.5 xhigh): caught 5+4+3+4 findings
  across the move list / visibility / wording cleanup.
- Gemini rounds 1-2: round-1 caught the substantive
  account_* coupling finding (false lockstep claim);
  round-2 returned PLAN-READY on v5.
